+++ ChangeLog.md
@@ -9,10 +9,12 @@ SPDX-License-Identifier: MIT
 <!-- Don't forget to update the gas/transaction costs tables in the
 README when a new version is released. -->
 
+## 1.5.0
+
 * [#144](https://github.com/tqtezos/stablecoin/pull/144)
   * Add `total_supply` field to the contract.
 
-* [142](https://github.com/tqtezos/stablecoin/pull/142)
+* [#142](https://github.com/tqtezos/stablecoin/pull/142)
   * Deploy TZIP-16 metadata to an external contract by default.
   * Provide deploy options `--contract-metadata-in-place` and
     `--contract-metadata-off-chain` to let user to embed the metadata